+++ config/routes.rb
@@ -2,7 +2,11 @@
 
 Rails.application.routes.draw do
   devise_for :users,
-             controllers: { omniauth_callbacks: "users/omniauth_callbacks" }
+             controllers: { omniauth_callbacks: "users/omniauth_callbacks" },
+             skip: [:sessions]
+  as :user do
+    delete "users/logout", to: "devise/sessions#destroy", as: :destroy_user_session
+  end
 
   resources :services, only: [:index, :show]